Add TutorTopicController

+++ src/main/java/com/devincubator/repository/QuestionRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
 
-    Question findByQuestionId(long questionId);
-    
+    Question findByQuestionId(int questionId);
+
 }
